requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-allauth==0.32.0
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-GNU-Terry-Pratchett==0.2.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.5
Markdown==2.6.8
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
PyJWT==1.5.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
urllib3==1.22
